Switch MethodNum and ActorID to aliases (#317)
* Switch MethodNum to an alias

* Switch ActorID to u64 alias

+++ vm/actor/src/builtin/multisig/mod.rs
@@ -18,6 +18,7 @@ use vm::{ActorError, ExitCode, MethodNum, Serialized, TokenAmount, METHOD_CONSTR
 
 /// Multisig actor methods available
 #[derive(FromPrimitive)]
+#[repr(u64)]
 pub enum Method {
     Constructor = METHOD_CONSTRUCTOR,
     Propose = 2,
@@ -34,7 +35,7 @@ pub enum Method {
 impl Method {
     /// Converts a method number into a Method enum
     fn from_method_num(m: MethodNum) -> Option<Method> {
-        FromPrimitive::from_u64(u64::from(m))
+        FromPrimitive::from_u64(m)
     }
 }
